Add files via upload
Instructions for how the code for this project is organized.
The raw data can be downloaded from Kaggle StateFarm competition page.   The data was extracted and organized as provided by Kaggle.
Note that we used Google Colaboratory Engine to build our models to take advantage of the free GPU provided.  So to use this code in it’s entirely as is, it must be used in Google Colab environment through Google Drive. 
However, one could still run the code in standalone computer, using Python or Ipython interpreter.  In such case, the code chunk that is used to connect to Google Colab should just be ignored or deleted.
The process followed for this code was as follows:
1.	The first file to use is create_train_val_sets_for_colab.py 
a.	This script reads loads the raw data and splits the data based on randomly selected train /validation set.
b.	The data can then be saved in pickle format to retrieve later or upload to google drive efficiently to be used with Colaboratory scripts.
c.	The data is saved with names that easily identify the valid ids set used to generate the train/validation data.  
d.	Note minor image processing is done as part of loading the data.

2.	Use any of the various model specific Ipython scripts to train and test models. There are scripts for:
a.	Logistic Regression
b.	Simple NNet and a Simple CNN (in one script)
c.	Deep CNN network
d.	PCA based models.

3.	All scripts are currently coded to read pickled data set generated by running the first script.  You simple have to matchup the valid ids set # to select the right data set to train with.
4.	All scripts provide options to save results and models using the same valid ids set based naming scheme for easy identification.